Trying to switch to Bootstrap 3 and make major internal changes
AT THE SAME TIME. So basically I had to give the thread database object
a normal number ID in addition to its reddit ID in order for the
internal Sequelize goodness to work. This means revising a large number
of my game logic to reflect this dual-ID system (which I'm sure won't be
confusing at all).

On the plus side, I'm mostly done switching to Bootstrap 3. This is all
for the best :^)